+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - No Nyaa results should have None for title anymore.
 - Distrowatch, Limetorrents and TPB should no longer throw error when returning no results.
 - Fix Etree search.
+- Fix Linuxtracker plugin.
 
 ## [1.2.0] - 2018-06-20
 ### Added